fix: address code review issues (C1, I1-I3, I5)
C1: Guard delta_mass min/max against Infinity when empty
I1: Defensive L0 fallback in L1 branch for exact-match-via-normalization
I2: Widen mismatch counter from u8 to u16 for future-proofing
I3: run_report placeholder now returns error instead of silent Ok
I5: Escape </script> sequences in JSON before HTML injection

+++ crates/entrapment-analysis/src/config.rs
@@ -118,7 +118,7 @@ pub struct SimilarityConfig {
     /// Maximum number of amino-acid mismatches allowed before considering
     /// two peptides distinct.
     #[serde(default = "default_max_mismatches")]
-    pub max_mismatches: u8,
+    pub max_mismatches: u16,
 
     /// Precursor m/z tolerance (Da) for linking similar peptides.
     #[serde(default = "default_delta_mz_threshold_da")]
@@ -133,7 +133,7 @@ pub struct SimilarityConfig {
     pub max_missed_cleavages: u32,
 }
 
-fn default_max_mismatches() -> u8 {
+fn default_max_mismatches() -> u16 {
     2
 }
 fn default_delta_mz_threshold_da() -> f64 {
